+++ python/CMakeLists.txt
@@ -44,9 +44,8 @@ install(PROGRAMS "${CMAKE_BINARY_DIR}/bin/revng" DESTINATION bin/)
 #
 # Install revng.model (including autogenerated classes)
 #
-set(PYTHON_MODEL_FILES
-    revng/model/__init__.py revng/model/_util.py revng/model/metaaddress.py
-    revng/model/v1/__init__.py revng/model/v1/external.py)
+set(PYTHON_MODEL_FILES revng/model/__init__.py revng/model/metaaddress.py
+                       revng/model/v1/__init__.py revng/model/v1/external.py)
 python_module(TARGET_NAME python-model MODULE_FILES ${PYTHON_MODEL_FILES}
               MODULE_GENERATED_FILES "${PYTHON_GENERATED_MODEL_PATH}")
