+++ python/revng/model/v1/__init__.py
@@ -4,9 +4,8 @@
 # flake8: noqa: F405
 # type: ignore
 
-from revng.tupletree import Reference, init_reference_yaml_classes
+from revng.tupletree import Reference, enum_value_to_index, init_reference_yaml_classes
 
-from .._util import enum_value_to_index
 from ..metaaddress import *
 from . import _generated
 from ._generated import *
